pyproject.toml
[project]
name = "testing-github-actions-self-hosted-runner-container"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Atsushi Morimoto (74th)", email = "74th.tech@gmail.com" }
]
dependencies = [
    "invoke>=2.2.0",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/testing_github_actions_self_hosted_runner_container"]
